requirements.txt
Pillow==9.1.1
pytube==12.1.0
pytube3==9.6.4
tk==0.1.0
typing_extensions==4.2.0
